More Windows fixes

+++ net/aeUuid.cpp
@@ -310,16 +310,16 @@ Guid GuidGenerator::newGuid()
 
 	const unsigned char bytes[16] = 
 	{
-		(newId.Data1 >> 24) & 0xFF,
-		(newId.Data1 >> 16) & 0xFF,
-		(newId.Data1 >> 8) & 0xFF,
-		(newId.Data1) & 0xff,
+		(uint8_t)( (newId.Data1 >> 24) & 0xFF ),
+		(uint8_t)( (newId.Data1 >> 16) & 0xFF ),
+		(uint8_t)( (newId.Data1 >> 8) & 0xFF ),
+		(uint8_t)( (newId.Data1) & 0xff ),
 
-		(newId.Data2 >> 8) & 0xFF,
-		(newId.Data2) & 0xff,
+		(uint8_t)( (newId.Data2 >> 8) & 0xFF ),
+		(uint8_t)( (newId.Data2) & 0xff ),
 
-		(newId.Data3 >> 8) & 0xFF,
-		(newId.Data3) & 0xFF,
+		(uint8_t)( (newId.Data3 >> 8) & 0xFF ),
+		(uint8_t)( (newId.Data3) & 0xFF ),
 
 		newId.Data4[0],
 		newId.Data4[1],
